--- Errors.py
@@ -1,8 +1,0 @@
-def view_error_rows():
-    pass
-
-
-def categorize_errors():
-    pass
-
-
