--- Hyperparameters.py
@@ -1,15 +1,0 @@
-lambda_expressions = [0, 0.01, 0.02, 0.04, 0.08, 0.16, 0.32, 0.64, 1.28, 2.58, 5.16, 10.24]
-alpha_values = [0.01, 0.05, 0.1, 0.25, 0.5, 0.75, 0.99]
-iterations = 200
-
-
-def get_lambda_expressions():
-    return lambda_expressions
-
-
-def get_alpha_values():
-    return alpha_values
-
-
-def get_iterations():
-    return iterations
