+++ csv_files/2.02. Binary predictors.csv
@@ -1,0 +1,169 @@
+SAT,Admitted,Gender
+1363,No,Male
+1792,Yes,Female
+1954,NA,Female
+1653,No,Male
+1593,No,Male
+1755,Yes,Female
+1775,Yes,Female
+1887,Yes,Female
+1893,Yes,Female
+5000,No,Male
+1857,Yes,Female
+1880,Yes,Female
+1664,Yes,Female
+1364,No,Male
+1693,No,Male
+1850,Yes,Female
+1633,No,Male
+1634,No,Male
+1636,Yes,Female
+1855,Yes,Female
+1987,Yes,Female
+1997,Yes,Male
+1422,No,Female
+1508,No,Female
+1720,Yes,Female
+1879,Yes,Male
+1634,Yes,Female
+1802,Yes,Male
+1849,Yes,Female
+1764,Yes,Female
+1460,No,Male
+1675,Yes,Female
+1656,No,Male
+2020,Yes,Male
+1850,Yes,Male
+1865,Yes,Male
+1664,No,Male
+1872,Yes,Female
+1654,No,Male
+1393,No,Female
+1587,No,Male
+1631,Yes,Female
+1931,Yes,Male
+1370,No,Male
+1810,Yes,Male
+1414,No,Male
+1761,Yes,Female
+1477,No,Male
+1486,No,Male
+1561,No,Male
+1549,No,Male
+2050,Yes,Male
+1697,No,Female
+1543,No,Female
+1934,Yes,Female
+1385,No,Male
+1670,No,Male
+1735,Yes,Female
+1634,No,Male
+1777,Yes,Female
+1550,No,Male
+1715,Yes,Male
+1925,Yes,Female
+1842,Yes,Male
+1786,Yes,Male
+1435,No,Male
+1387,No,Male
+1521,No,Male
+1975,Yes,Female
+1435,No,Male
+1714,Yes,Male
+1634,Yes,Male
+1464,No,Female
+1794,Yes,Male
+1855,Yes,Male
+1953,Yes,Male
+1469,No,Male
+1663,Yes,Male
+1907,Yes,Female
+1990,Yes,Female
+1542,No,Male
+1808,Yes,Female
+1966,Yes,Female
+1679,No,Female
+2021,Yes,Female
+2015,Yes,Female
+1473,No,Male
+1979,Yes,Male
+1787,Yes,Female
+1687,Yes,Female
+1674,No,Female
+1478,No,Female
+1735,Yes,Female
+1720,Yes,Female
+1494,No,Male
+1964,Yes,Male
+1843,Yes,Female
+1550,No,Female
+1764,Yes,Female
+1712,Yes,Female
+1775,Yes,Female
+1531,No,Male
+1781,Yes,Female
+1579,No,Male
+1526,No,Female
+1778,Yes,Female
+1769,Yes,Female
+1824,Yes,Female
+1481,No,Male
+1464,No,Male
+1591,No,Male
+1666,No,Male
+1455,No,Male
+1934,Yes,Female
+1625,No,Male
+1334,No,Male
+1721,Yes,Female
+1475,No,Male
+1662,Yes,Female
+1861,Yes,Female
+1936,Yes,Female
+1572,No,Male
+1508,No,Female
+1430,No,Male
+1891,Yes,Female
+1550,No,Male
+1741,Yes,Female
+1690,No,Male
+1687,Yes,Male
+1730,Yes,Male
+1674,Yes,Female
+1475,No,Male
+1962,Yes,Female
+1532,No,Male
+1492,No,Male
+1502,No,Male
+1974,Yes,Female
+1607,No,Male
+1412,No,Female
+1557,No,Male
+1821,Yes,Male
+1760,Yes,Female
+1685,Yes,Male
+1773,Yes,Female
+1826,Yes,Female
+1565,No,Female
+1510,No,Male
+1374,No,Male
+1402,No,Male
+1702,Yes,Male
+1956,Yes,Female
+1933,Yes,Male
+1832,Yes,Male
+1893,Yes,Male
+1831,Yes,Female
+1487,No,Male
+2041,Yes,Female
+1850,Yes,Male
+1555,No,Female
+2020,Yes,Female
+1593,No,Male
+1934,Yes,Female
+1808,Yes,Male
+1722,Yes,Female
+7000,Yes,Male
+1555,No,Male
+1524,No,Male
+1461,No,Male
