+++ learning_algorithms/support_algorithms/Fit.py
@@ -1,0 +1,23 @@
+
+
+class Fit:
+    def __init__(self):
+        pass
+
+    def overfitting(self):
+        # High Variance
+        def get_more_training_examples():
+            pass
+        def reduce_feature_set():
+            pass
+        def increasing_lambda():
+            pass
+
+    def underfitting(self):
+        # High Bias
+        def get_more_features():
+            pass
+        def add_polynomial():
+            pass
+        def decreasing_lambda():
+            pass
