+++ preprocess/ReformatDataFrame.py
@@ -1,0 +1,11 @@
+import pandas as pd
+import numpy as np
+
+
+def delete_row(train_data, row):
+    pass
+
+
+def delete_column(train_data, column):
+    return train_data.drop([column], axis=1)
+
